pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cli-agent"
version = "0.1.0"
description = "CLI AI Coding Agent"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "google-genai",
    "rich",
    "python-dotenv",
]

[project.scripts]
cli-agent = "cli_agent.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cli_agent*"]